backend/requirements.txt
asgiref==3.7.2
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.12.4
psycopg2==2.9.9
psycopg2-binary==2.9.3
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
Pillow==9.3.0